Add YAML configuration for AArch64 relocations in nn.dSYM

+++ Neural_Network_in_C/matrix/matrix.h
@@ -7,7 +7,7 @@ typedef struct {
     int cols;
 } Matrix;
 Matrix* matrix_create(int rows, int cols);
-void matrix_init(Matrix* matrix, int n);
+void matrix_init(Matrix* matrix, double n);
 void matrix_free(Matrix* matrix);
 void matrix_print(Matrix* matrix);
 Matrix* matrix_copy(Matrix* matrix);
